requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bleach==6.1.0
blessed==1.20.0
boto3==1.34.23
botocore==1.34.23
cachetools==5.3.2
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
colorama==0.4.6
colorlog==6.8.0
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==4.4.2
distlib==0.3.8
docker-pycreds==0.4.0
ensemble-boxes==1.0.9
exceptiongroup==1.2.0
executing==2.0.1
fast_pytorch_kmeans==0.1.6
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gpustat==1.1.1
grpcio==1.60.1
h5py==3.10.0
huggingface-hub==0.20.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
icecream==2.1.3
idna==3.6
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kaggle==1.6.3
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.10.1
lit==17.0.6
llvmlite==0.46.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.9
networkx==3.2.1
numba==0.63.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.1.0
proglog==0.1.12
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyDeprecate==0.3.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-lightning==2.0.8
pytorch-pretrained-bert==0.6.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
rtpt==0.0.4
s3transfer==0.10.0
safetensors==0.4.1
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.11.4
sentry-sdk==2.47.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
SoundFile==0.10.3.post1
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2025.5.10
timm==0.9.7
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.3.0.post0
torchvision==0.15.2
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
triton==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
urllib3==1.26.18
virtualenv==20.25.0
Wand==0.6.13
wandb==0.23.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
